requirements.txt
fastapi
uvicorn
pymongo
chromadb
openai
requests
python-dotenv
duckduckgo_search
